Line Detection Algorithm

+++ src/lineSensor.cpp
@@ -11,14 +11,13 @@ int adcpin6 = 21;
    adc5.begin(adcpin5,11,12,13);
    adc6.begin(adcpin6,11,12,13);
 
-   angleLS = new double[24];
+   angleLS = new int[24];
 	for(int i = 0; i < 24; i ++) {
 		angleLS[i] = cos(toRadians( i * 15));
-	
 	}
 };
 
-double *LineSensor::LineAngle()
+int *LineSensor::LineAngle()
 {
     return angleLS;
 }
